got modal working on create new feed
