+++ react-app/src/components/AddFeedForm/index.js
@@ -1,47 +1,20 @@
-import React, { useState } from "react";
-import { useHistory } from 'react-router-dom';
-import { useDispatch } from "react-redux";
-import { postFeed } from '../../store/feeds'
-import './AddFeedForm.css';
+import React, { useState } from 'react';
+import { Modal } from '../../context/Modal';
+import AddFeedForm from './AddFeedForm.js';
 
-function AddFeedForm() {
-    const dispatch = useDispatch()
-    const history = useHistory();
-    const [feedName, setFeedName] = useState('');
+function CreateFeedModal() {
+  const [showModal, setShowModal] = useState(false);
 
-    const handleSubmit = async (e) => {
-        e.preventDefault();
-
-        const payload = {
-            feedName
-        }
-
-        await dispatch(postFeed(payload));
-        setFeedName('');
-
-        history.push('/');
-    }
-
-    return (
-        <div className="form-container">
-            <form onSubmit={handleSubmit}>
-                <div className="form__input-fields">
-                    <label>Title</label>
-                    <input
-                        type="text"
-                        value={feedName}
-                        onChange={(e) => (setFeedName(e.target.value))}
-                        required
-                        placeholder="Feed Name"
-                    />
-                </div>
-                <div className="form__btn">
-                    <button type="submit" id="signup-btn">Save</button>
-                </div>
-            </form>
-        </div>
-    )
+  return (
+    <>
+      <button onClick={() => setShowModal(true)}>Create a new feed</button>
+      {showModal && (
+        <Modal onClose={() => setShowModal(false)}>
+          <AddFeedForm />
+        </Modal>
+      )}
+    </>
+  );
 }
 
-
-export default AddFeedForm
+export default CreateFeedModal;
