+++ react-app/src/index.js
@@ -5,18 +5,18 @@ import './index.css';
 import App from './App';
 import { BrowserRouter } from 'react-router-dom';
 import configureStore from "./store"
-// import { ModalProvider } from './context/Modal'
+import { ModalProvider } from './context/Modal'
 
 const store = configureStore()
 
 function Root() {
     return (
         <Provider store={store}>
-            {/* <ModalProvider> */}
-            <BrowserRouter>
-                <App />
-            </BrowserRouter>
-            {/* </ModalProvider> */}
+            <ModalProvider>
+                <BrowserRouter>
+                    <App />
+                </BrowserRouter>
+            </ModalProvider>
         </Provider>
     )
 }
